+++ packages/agent-library/test/_mocks_/mock-models.ts
@@ -1,0 +1,17 @@
+import {
+  type AgentInvokeOptions,
+  type AgentProcessResult,
+  ChatModel,
+  type ChatModelInput,
+  type ChatModelOutput,
+} from "@aigne/core";
+import type { PromiseOrValue } from "@aigne/core/utils/type-utils.js";
+
+export class OpenAIChatModel extends ChatModel {
+  process(
+    _input: ChatModelInput,
+    _options: AgentInvokeOptions,
+  ): PromiseOrValue<AgentProcessResult<ChatModelOutput>> {
+    throw new Error("Method not implemented.");
+  }
+}