+++ packages/transport/typedoc.js
@@ -1,0 +1,10 @@
+import base from "../../typedoc.base.js";
+
+/**
+ * @type {import('typedoc').TypeDocOptions}
+ */
+export default {
+  ...base,
+  entryPoints: ["src/http-server/index.ts", "src/http-client/index.ts"],
+  readme: "README.md",
+};